requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mypysql==0.1.7
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.1
numpy==1.26.3
pandas==2.1.4
psycopg2==2.9.9
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
Werkzeug==3.0.1
zipp==3.17.0